Added working with closing
