+++ extensions/vscode/src/ContinueGUIWebviewViewProvider.ts
@@ -83,7 +83,7 @@ export class ContinueGUIWebviewViewProvider
 
   public resetWebviewProtocolWebview(): void {
     if (this._webview) {
-      this.webviewProtocol.resetWebviews()
+      this.webviewProtocol.resetWebviewToDefault()
       // this.webviewProtocol.addWebview(this._webview);
     } else {
       console.warn("no webview found during reset");
